I only changed the MyAccountWindow
The other ones are either insignificant changes or deleted empty rows.
So now we can see the details of the user thanks to repeated code but I
couldn't figure out how to do it without bumping into
NullPointerException because we have 2 kind of users with the same
attributes so.... same same but different different.