Fully functional view and edit details;
Register changed, included static checker for type from the checkbox
directly and got of the second User object.
Also I've deleted one constructor and one query from the User class.
Now when the checkbox is not selected(normal customer) the fields from
the corporate part of the user(which are not visible) are empty Strings.
For the actual corporate users it's alright from the beginning.
Also MyAccountWindow has no repeated code so I hope you will make peace
with the exceptions which are there only because data was inserted
manually :)).