Booking class built, NewBookingView complete
NewBooking now has the scrollPane to display bands and is actually writing to the db.
Fun features: Made the links work! they open in the default browser!
Msg before booking changes depanding on users payment method
Button is inactive if the user has picked no tickets, or if there aren't enough available.

On to Booking History :)